requirements.txt
flask==3.1.1
python-dotenv==1.1.0
openpyxl==3.1.5
psycopg2-binary==2.9.10
gunicorn==23.0.0
pandas==2.2.3
numpy==2.2.4
flask-login==0.6.3
bcrypt==4.2.1
pyotp==2.9.0
qrcode[pil]==8.0
flask-limiter==3.8.0
rapidfuzz==3.10.1
